gandalf/pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.backends._legacy:_Backend"

[project]
name = "sleep"
version = "0.1.0"
description = "Sleep staging from EEG using deep learning"
requires-python = ">=3.10"
dependencies = [
    "torch",
    "numpy",
    "pandas",
    "matplotlib",
    "seaborn",
    "scikit-learn",
    "mne",
    "pyyaml",
    "tqdm",
    "tensorboard",
    "ipykernel",
]

[project.scripts]
train-gandalf = "scripts.train_gandalf:main"
eval-gandalf = "scripts.eval_gandalf:main"

[tool.setuptools.packages.find]
where = ["src"]
